+++ BlaBlaMessengerServer/src/data_structures/CommandData.java
@@ -1,7 +1,6 @@
 package data_structures;
 
 import java.io.Serializable;
-import java.util.UUID;
 
 public class CommandData implements Serializable {
     public Commands Command;
@@ -15,11 +14,5 @@ public class CommandData implements Serializable {
         Command = command;
         Id = id;
         Data = data;
-    }
-    
-    public class CommandId extends DataObject {
-        public UUID Id;
-    
-        public CommandId( final UUID id ) { Id = id; }
     }
 }
