+++ BlaBlaMessengerServer/src/data_structures/CommandId.java
@@ -1,0 +1,9 @@
+package data_structures;
+
+import java.util.UUID;
+
+public class CommandId extends DataObject {
+    public UUID Id;
+
+    public CommandId( final UUID id ) { Id = id; }
+}
