+++ BlaBlaMessengerServer/src/data_structures/ConferenceId.java
@@ -1,0 +1,9 @@
+package data_structures;
+
+import java.util.UUID;
+
+public class ConferenceId extends DataObject {
+    public UUID Id;
+
+    public ConferenceId() { Id = UUID.randomUUID(); }
+}
