+++ BlaBlaMessengerServer/src/data_structures/ConferenceName.java
@@ -1,0 +1,7 @@
+package data_structures;
+
+public class ConferenceName extends DataObject {
+    public String Name;
+
+    public ConferenceName( String name ) { Name = name; }
+}
