+++ BlaBlaMessengerServer/src/data_structures/ContactConfPair.java
@@ -1,8 +1,5 @@
 package data_structures;
 
-import data_structures.Contact.ContactId;
-import data_structures.Conference.ConferenceId;
-
 public class ContactConfPair extends DataObject {
     public ContactId Contact;
     public ConferenceId Conference;
