+++ BlaBlaMessengerServer/src/data_structures/ContactName.java
@@ -1,0 +1,7 @@
+package data_structures;
+
+public class ContactName extends DataObject {
+    public String Name;
+
+    public ContactName( final String name  ) { Name = name; }
+}
