+++ BlaBlaMessengerServer/src/data_structures/FileData.java
@@ -1,0 +1,7 @@
+package data_structures;
+
+public class FileData extends DataObject {
+    public byte[] Data;
+
+    public FileData( final byte[] data ) { Data = data; }
+}
