+++ BlaBlaMessengerServer/src/data_structures/FileId.java
@@ -1,0 +1,9 @@
+package data_structures;
+
+import java.util.UUID;
+
+public class FileId extends DataObject {
+    public UUID Id;
+
+    public FileId() { Id = UUID.randomUUID(); }
+}
