+++ BlaBlaMessengerServer/src/data_structures/FileIdNamePair.java
@@ -1,17 +1,6 @@
 package data_structures;
 
-import data_structures.DataObject;
-import data_structures.File.FileName;
-import java.util.UUID;
-
 public class FileIdNamePair extends DataObject {
-    public class FileId extends DataObject {
-        public UUID Id;
-    
-        public FileId() { Id = UUID.randomUUID(); }
-    }
-
-    
     public FileId Id;
     public FileName Name;
