+++ BlaBlaMessengerServer/src/data_structures/FileIdNamePairs.java
@@ -5,6 +5,6 @@ import java.util.ArrayList;
 public class FileIdNamePairs extends DataObject {
     public ArrayList< FileIdNamePair > Pairs;
     
-    public FileIdNamePairs( ArrayList< FileIdNamePair > pairs )
+    public FileIdNamePairs( ArrayList< FileIdNamePair > pairs ) 
     { Pairs = pairs; }
 }
