+++ BlaBlaMessengerServer/src/data_structures/FileName.java
@@ -1,0 +1,7 @@
+package data_structures;
+
+public class FileName extends DataObject {
+    public String Name;
+
+    public FileName( String name ) { Name = name; }
+}
